+++ src/grepros/__init__.py
@@ -6,10 +6,10 @@ Released under the BSD License.
 
 @author      Erki Suurjaak
 @created     31.10.2021
-@modified    10.02.2022
+@modified    11.02.2022
 ------------------------------------------------------------------------------
 """
 ## @namespace grepros
 __title__        = "grepros"
-__version__      = "0.4.3.dev2"
-__version_info__ = (0, 4, 3, "dev2")
+__version__      = "0.4.3.dev3"
+__version_info__ = (0, 4, 3, "dev3")